Add files via upload
I'm sorry it was late I had an extreme amount of trouble with this lab I have spent various hours trying to figure out what I did wrong. I figured out most of it on my own and emailed you for help as well. The last section was a little difficult since I had to work off my desktop since the raw_data wouldn't transfer to my documents. I am so sorry about the late work, it will not happen again, I was just having a lot of difficulty. Please let me know if it needs to be updated or fixed if I went wrong anywhere.